+++ cf/test/individual_tests.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 
-file=create_test_files.py
-echo "Running $file"
-python $file
-rc=$?
-if [[ $rc != 0 ]]; then
-  exit $rc
-fi
+for file in create_test_files*.py
+do
+  echo "Running $file"
+  python $file
+  rc=$?
+  if [[ $rc != 0 ]]; then
+    exit $rc
+  fi
+done
 
 file=setup_create_field.py
 echo "Running $file"